midi presets bank added
add:
* static starter general midi presets bank
fix:
* dac pan writing at place of instrument data (first few instruments was corrupted)

+++ src/MDT_PCM.h
@@ -8,7 +8,7 @@
 static u32 SRAM_ReadSampleRegion(u8 bank, u8 note, u8 byteNum);
 void SRAM_WriteSampleRegion(u8 bank, u8 note, u8 byteNum, u8 data);
 
-static u32 SRAM_ReadSamplePan(u8 bank, u8 note);
+static u8 SRAM_ReadSamplePan(u8 bank, u8 note);
 void SRAM_WriteSamplePan(u8 bank, u8 note, u8 data);
 
 #endif // MDT_PCM_H_INCLUDED
